+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Simplify logic which determines whether ackOrNack should be a callback or a promise
+- Report repeated calls to ackOrNack
 
 ## 16.2.0